pyproject.toml
[tool.poetry]
name = "nitbot"
version = "1.0.0"
description = "A simple Discord bot to tell you if your users' deranged ramblings are holy or not"
authors = ["Brett Heinkel <bheinks@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
discord-py = "^2.2.2"
python-decouple = "^3.8"
nltk = "^3.8.1"
contractions = "^0.1.73"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
